Order Service Endpoints & config server

+++ EcomMicroservices/productService/src/main/java/com/cloudthat/productService/model/ProductModel.java
@@ -8,7 +8,7 @@ public class ProductModel {
     private Long id;
     private String name;
     private  Double price;
-    private Integer quantity;
+    private Long quantity;
     private Category category;
 
     public String getName() {
@@ -27,11 +27,11 @@ public class ProductModel {
         this.price = price;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
 
@@ -51,14 +51,14 @@ public class ProductModel {
         this.category = category;
     }
 
-    public ProductModel(String name, Double price, Integer quantity, Category category) {
+    public ProductModel(String name, Double price, Long quantity, Category category) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
         this.category = category;
     }
 
-    public ProductModel(Long id, String name, Double price, Integer quantity, Category category) {
+    public ProductModel(Long id, String name, Double price, Long quantity, Category category) {
         this.id = id;
         this.name = name;
         this.price = price;
